requirements.txt
django
python-telegram-bot
django-apscheduler
python-dotenv